+++ nautilus_trader/backtest/matching_engine.pyx
@@ -184,7 +184,7 @@ cdef class OrderMatchingEngine:
         self.book_type = book_type
         self.oms_type = oms_type
         self.account_type = account_type
-        self.market_status = MarketStatus.OPEN
+        self.market_status = MarketStatusAction.TRADING
 
         self._instrument_has_expiration = instrument.instrument_class in EXPIRING_INSTRUMENT_TYPES
         self._bar_execution = bar_execution
@@ -519,36 +519,39 @@ cdef class OrderMatchingEngine:
                 f"invalid `PriceType`, was {price_type}",  # pragma: no cover
             )
 
-    cpdef void process_status(self, MarketStatus status):
+    cpdef void process_status(self, MarketStatusAction status):
         """
         Process the exchange status.
 
         Parameters
         ----------
-        status : MarketStatus
-            The status to process.
+        status : MarketStatusAction
+            The status action to process.
 
         """
-        if (self.market_status, status) == (MarketStatus.CLOSED, MarketStatus.OPEN):
-            self.market_status = status
-        elif (self.market_status, status) == (MarketStatus.CLOSED, MarketStatus.PRE_OPEN):
-            # Nothing to do on pre-market open.
-            self.market_status = status
-        elif (self.market_status, status) == (MarketStatus.PRE_OPEN, MarketStatus.PAUSE):
-            # Opening auction period, run auction match on pre-open auction orderbook
-            self.process_auction_book(self._opening_auction_book)
-            self.market_status = status
-        elif (self.market_status, status) == (MarketStatus.PAUSE, MarketStatus.OPEN):
-            # Normal market open
-            self.market_status = status
-        elif (self.market_status, status) == (MarketStatus.OPEN, MarketStatus.PAUSE):
-            # Closing auction period, run auction match on closing auction orderbook
-            self.process_auction_book(self._closing_auction_book)
-            self.market_status = status
-        elif (self.market_status, status) == (MarketStatus.PAUSE, MarketStatus.CLOSED):
-            # Market closed - nothing to do for now
-            # TODO - should we implement some sort of closing price message here?
-            self.market_status = status
+        self.market_status = status
+
+        # TODO: Reimplement
+        # if (self.market_status, status) == (MarketStatusAction.CLOSED, MarketStatus.OPEN):
+        #     self.market_status = status
+        # elif (self.market_status, status) == (MarketStatus.CLOSED, MarketStatus.PRE_OPEN):
+        #     # Nothing to do on pre-market open.
+        #     self.market_status = status
+        # elif (self.market_status, status) == (MarketStatus.PRE_OPEN, MarketStatus.PAUSE):
+        #     # Opening auction period, run auction match on pre-open auction orderbook
+        #     self.process_auction_book(self._opening_auction_book)
+        #     self.market_status = status
+        # elif (self.market_status, status) == (MarketStatus.PAUSE, MarketStatus.OPEN):
+        #     # Normal market open
+        #     self.market_status = status
+        # elif (self.market_status, status) == (MarketStatus.OPEN, MarketStatus.PAUSE):
+        #     # Closing auction period, run auction match on closing auction orderbook
+        #     self.process_auction_book(self._closing_auction_book)
+        #     self.market_status = status
+        # elif (self.market_status, status) == (MarketStatus.PAUSE, MarketStatus.CLOSED):
+        #     # Market closed - nothing to do for now
+        #     # TODO - should we implement some sort of closing price message here?
+        #     self.market_status = status
 
     cpdef void process_auction_book(self, OrderBook book):
         Condition.not_none(book, "book")