+++ nautilus_trader/core/rust/model.pyx
@@ -23,11 +23,10 @@ from nautilus_trader.core.rust.model cimport BookAction  # type: ignore
 from nautilus_trader.core.rust.model cimport BookType  # type: ignore
 from nautilus_trader.core.rust.model cimport ContingencyType  # type: ignore
 from nautilus_trader.core.rust.model cimport CurrencyType  # type: ignore
-from nautilus_trader.core.rust.model cimport HaltReason  # type: ignore
 from nautilus_trader.core.rust.model cimport InstrumentClass  # type: ignore
 from nautilus_trader.core.rust.model cimport InstrumentCloseType  # type: ignore
 from nautilus_trader.core.rust.model cimport LiquiditySide  # type: ignore
-from nautilus_trader.core.rust.model cimport MarketStatus  # type: ignore
+from nautilus_trader.core.rust.model cimport MarketStatusAction  # type: ignore
 from nautilus_trader.core.rust.model cimport OmsType  # type: ignore
 from nautilus_trader.core.rust.model cimport OptionKind  # type: ignore
 from nautilus_trader.core.rust.model cimport OrderSide  # type: ignore