+++ tests/integration_tests/adapters/betfair/test_betfair_parsing.py
@@ -83,7 +83,7 @@ from nautilus_trader.model.data import InstrumentStatus
 from nautilus_trader.model.data import OrderBookDeltas
 from nautilus_trader.model.data import TradeTick
 from nautilus_trader.model.enums import AccountType
-from nautilus_trader.model.enums import MarketStatus
+from nautilus_trader.model.enums import MarketStatusAction
 from nautilus_trader.model.enums import OrderSide
 from nautilus_trader.model.enums import OrderStatus as NautilusOrderStatus
 from nautilus_trader.model.enums import TimeInForce
@@ -134,7 +134,7 @@ class TestBetfairParsingStreaming:
         result = [
             upd
             for upd in updates
-            if isinstance(upd, InstrumentStatus) and upd.status == MarketStatus.PRE_OPEN
+            if isinstance(upd, InstrumentStatus) and upd.action == MarketStatusAction.PRE_OPEN
         ]
         assert len(result) == 17